Move cipher utility functions to a new file
